requirements.txt
streamlit>=1.37.0
pandas
fpdf2